+++ src/Git_Updater/API/API.php
@@ -333,7 +333,7 @@ class API {
 		$response = isset( $this->response['dot_org'] ) ? $this->response['dot_org'] : false;
 
 		if ( ! $response ) {
-			$url      = "https://api.wordpress.org/{$this->type->type}s/info/1.1/";
+			$url      = "https://api.wordpress.org/{$this->type->type}s/info/1.2/";
 			$url      = add_query_arg(
 				[
 					'action'                        => "{$this->type->type}_information",
@@ -341,7 +341,7 @@ class API {
 				],
 				$url
 			);
-			$response = wp_remote_get( $url );
+			$response = wp_remote_head( $url );
 
 			if ( is_wp_error( $response ) ) {
 				Singleton::get_instance( 'Messages', $this )->create_error_message( $response );
@@ -349,8 +349,8 @@ class API {
 				return false;
 			}
 
-			$response = json_decode( $response['body'] );
-			$response = ! empty( $response ) && ! isset( $response->error ) ? 'in dot org' : 'not in dot org';
+			$code     = wp_remote_retrieve_response_code( $response );
+			$response = 200 === $code ? 'in dot org' : 'not in dot org';
 
 			$this->set_repo_cache( 'dot_org', $response );
 		}
@@ -495,32 +495,8 @@ class API {
 	 * @access protected
 	 */
 	protected function add_meta_repo_object() {
-		$this->type->rating       = $this->make_rating( $this->type->repo_meta );
 		$this->type->last_updated = $this->type->repo_meta['last_updated'];
-		$this->type->num_ratings  = $this->type->repo_meta['watchers'];
 		$this->type->is_private   = $this->type->repo_meta['private'];
-	}
-
-	/**
-	 * Create some sort of rating from 0 to 100 for use in star ratings.
-	 * I'm really just making this up, more based upon popularity.
-	 *
-	 * @param array $repo_meta Array of repo meta data.
-	 *
-	 * @return integer
-	 */
-	protected function make_rating( $repo_meta ) {
-		$watchers    = ! empty( $repo_meta['watchers'] ) ? $repo_meta['watchers'] : 0;
-		$forks       = ! empty( $repo_meta['forks'] ) ? $repo_meta['forks'] : 0;
-		$open_issues = ! empty( $repo_meta['open_issues'] ) ? $repo_meta['open_issues'] : 0;
-
-		$rating = abs( (int) round( $watchers + ( $forks * 1.5 ) - ( $open_issues * 0.1 ) ) );
-
-		if ( 100 < $rating ) {
-			return 100;
-		}
-
-		return $rating;
 	}
 
 	/**